requirements.txt
click==8.1.2
colorama==0.4.4
Flask==2.1.1
fpdf==1.7.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
python-dateutil==2.8.2
six==1.16.0
Werkzeug==2.1.1
zipp==3.7.0
